blocksync: Move into internal
