+++ internal/blocksync/msgs_test.go
@@ -9,7 +9,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"github.com/cometbft/cometbft/blocksync"
+	"github.com/cometbft/cometbft/internal/blocksync"
 	bcproto "github.com/cometbft/cometbft/proto/tendermint/blocksync"
 	"github.com/cometbft/cometbft/types"
 )
@@ -93,26 +93,44 @@ func TestBlocksyncMessageVectors(t *testing.T) {
 		expBytes string
 	}{
 		{"BlockRequestMessage", &bcproto.Message{Sum: &bcproto.Message_BlockRequest{
-			BlockRequest: &bcproto.BlockRequest{Height: 1}}}, "0a020801"},
-		{"BlockRequestMessage", &bcproto.Message{Sum: &bcproto.Message_BlockRequest{
-			BlockRequest: &bcproto.BlockRequest{Height: math.MaxInt64}}},
-			"0a0a08ffffffffffffffff7f"},
+			BlockRequest: &bcproto.BlockRequest{Height: 1},
+		}}, "0a020801"},
+		{
+			"BlockRequestMessage", &bcproto.Message{Sum: &bcproto.Message_BlockRequest{
+				BlockRequest: &bcproto.BlockRequest{Height: math.MaxInt64},
+			}},
+			"0a0a08ffffffffffffffff7f",
+		},
 		{"BlockResponseMessage", &bcproto.Message{Sum: &bcproto.Message_BlockResponse{
-			BlockResponse: &bcproto.BlockResponse{Block: bpb}}}, "1a700a6e0a5b0a02080b1803220b088092b8c398feffffff012a0212003a20c4da88e876062aa1543400d50d0eaa0dac88096057949cfb7bca7f3a48c04bf96a20e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855120d0a0b48656c6c6f20576f726c641a00"},
+			BlockResponse: &bcproto.BlockResponse{Block: bpb},
+		}}, "1a700a6e0a5b0a02080b1803220b088092b8c398feffffff012a0212003a20c4da88e876062aa1543400d50d0eaa0dac88096057949cfb7bca7f3a48c04bf96a20e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855120d0a0b48656c6c6f20576f726c641a00"},
 		{"NoBlockResponseMessage", &bcproto.Message{Sum: &bcproto.Message_NoBlockResponse{
-			NoBlockResponse: &bcproto.NoBlockResponse{Height: 1}}}, "12020801"},
-		{"NoBlockResponseMessage", &bcproto.Message{Sum: &bcproto.Message_NoBlockResponse{
-			NoBlockResponse: &bcproto.NoBlockResponse{Height: math.MaxInt64}}},
-			"120a08ffffffffffffffff7f"},
-		{"StatusRequestMessage", &bcproto.Message{Sum: &bcproto.Message_StatusRequest{
-			StatusRequest: &bcproto.StatusRequest{}}},
-			"2200"},
-		{"StatusResponseMessage", &bcproto.Message{Sum: &bcproto.Message_StatusResponse{
-			StatusResponse: &bcproto.StatusResponse{Height: 1, Base: 2}}},
-			"2a0408011002"},
-		{"StatusResponseMessage", &bcproto.Message{Sum: &bcproto.Message_StatusResponse{
-			StatusResponse: &bcproto.StatusResponse{Height: math.MaxInt64, Base: math.MaxInt64}}},
-			"2a1408ffffffffffffffff7f10ffffffffffffffff7f"},
+			NoBlockResponse: &bcproto.NoBlockResponse{Height: 1},
+		}}, "12020801"},
+		{
+			"NoBlockResponseMessage", &bcproto.Message{Sum: &bcproto.Message_NoBlockResponse{
+				NoBlockResponse: &bcproto.NoBlockResponse{Height: math.MaxInt64},
+			}},
+			"120a08ffffffffffffffff7f",
+		},
+		{
+			"StatusRequestMessage", &bcproto.Message{Sum: &bcproto.Message_StatusRequest{
+				StatusRequest: &bcproto.StatusRequest{},
+			}},
+			"2200",
+		},
+		{
+			"StatusResponseMessage", &bcproto.Message{Sum: &bcproto.Message_StatusResponse{
+				StatusResponse: &bcproto.StatusResponse{Height: 1, Base: 2},
+			}},
+			"2a0408011002",
+		},
+		{
+			"StatusResponseMessage", &bcproto.Message{Sum: &bcproto.Message_StatusResponse{
+				StatusResponse: &bcproto.StatusResponse{Height: math.MaxInt64, Base: math.MaxInt64},
+			}},
+			"2a1408ffffffffffffffff7f10ffffffffffffffff7f",
+		},
 	}
 
 	for _, tc := range testCases {
